fix: 보스 매칭 취소 시, 안내창 변경
