+++ Assets/Scripts/Boss/BossUIMatchingFail.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 534c72dfa42250f418338f0a8d2712f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
