--- Assets/Scripts/Boss/UIBossMatching.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 482d2f09902fe8d418b28f41d555fe2f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
